% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cutadapt_tibble}
\alias{cutadapt_tibble}
\title{Cutadapt Tibble}
\usage{
cutadapt_tibble(fastq_data, metadata)
}
\arguments{
\item{metadata}{The metadata table that prepare_metadata provides}

\item{new_fastq_data}{The fastq data that came out of prepare fastq}
}
\value{
A tibble for cutadapt
}
\description{
Creates a tibble that can be fed into cutadapt by combining metadata, primer, and fastq tibbles
}
