% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_and_trim}
\alias{filter_and_trim}
\title{Filter and Trim}
\usage{
filter_and_trim(intermediate_path, fastq_data)
}
\arguments{
\item{intermediate_path}{The path to the intermediate data file}

\item{fastq_data}{The tibble that the read_fastq function spits out}
}
\value{
None
}
\description{
To filter reads after cutadapt with various choices - i.e. with Ns, below a length of 50, or greater than expected error limit
}
