% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{matching_order_primer_check}
\alias{matching_order_primer_check}
\title{Mathcing Order Primer Check}
\usage{
matching_order_primer_check(fastq_data)
}
\arguments{
\item{fastq_data}{The tibble that the read_fastq function spits out}
}
\value{
None
}
\description{
Validation for the primers - making sure the forward and reverse reads are in matching order
}
