% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{post_primer_hit_data}
\alias{post_primer_hit_data}
\title{Post Primer Hit Data}
\usage{
post_primer_hit_data(
  primer_hit_data,
  primer_data,
  fastq_data,
  intermediate_path
)
}
\arguments{
\item{primer_hit_data}{the hit data from pre_primer_hit data}

\item{primer_data}{The tibble from prepare_primers}

\item{fastq_data}{The tibble that the read_fastq function spits out}

\item{intermediate_path}{The path to the intermediate data file}
}
\value{
A tibble with all the primer hits
}
\description{
Makes a primer hit tibble for fastq sequences AFTER cutadapt has been run
}
