% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_fastq}
\alias{prepare_fastq}
\title{Prepare Fastq}
\usage{
prepare_fastq(fastq_path, intermediate_path)
}
\arguments{
\item{fastq_path}{The path to the fastq files}

\item{intermediate_path}{The path to the intermediate data file}
}
\value{
A data frame with the fastq information
}
\description{
Does all of the processing/editing of the fastqs in order to make ready for cutadapt
}
