% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_primers}
\alias{prepare_primers}
\title{Prepare Primers}
\usage{
prepare_primers(primer_path)
}
\arguments{
\item{primer_path}{a path to the csv that holds the primer info for this project.}
}
\value{
A tibble that contains the forward and reverse primers with created complement, etc.
}
\description{
Take in user primers and creates the complement, reverse, reverse complement of primers in one tibble
}
