% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_ns}
\alias{remove_ns}
\title{Remove N's}
\usage{
remove_ns(fastq_data, intermediate_path)
}
\arguments{
\item{fastq_data}{The tibble that the read_fastq function spits out}

\item{intermediate_path}{The path to the intermediate data file}
}
\value{
A tibble with the sequences with the N's removed
}
\description{
Removes any sequences that have Ns in them
}
